requirements.txt
typer
pytest
requests
python-dotenv